+++ src/features/comment/components/index.ts
@@ -1,0 +1,1 @@
+export * from "./Comments"
