+++ src/features/comment/models/index.ts
@@ -1,0 +1,1 @@
+export * from "./queries"
