+++ src/features/post/components/index.ts
@@ -1,0 +1,5 @@
+export * from "./PostDialog"
+export * from "./PostCreateForm"
+export * from "./PostEditForm"
+export * from "./PostViewContent"
+export * from "./PostTable"
