+++ src/features/post/index.ts
@@ -1,4 +1,4 @@
+export * from "./apis"
 export * from "./hooks"
 export * from "./models"
-export * from "./ui"
-export * from "./apis"
+export * from "./components"
