+++ src/features/post/models/index.ts
@@ -1,1 +1,2 @@
 export * from "./types"
+export * from "./queries"
