+++ src/features/user/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from "./useUsersQuery"
 export * from "./useUserQuery"
+export * from "./useUserId"
