+++ src/features/user/index.ts
@@ -1,2 +1,3 @@
 export * from "./hooks"
 export * from "./models"
+export * from "./components"
