+++ src/shared/components/index.ts
@@ -6,3 +6,4 @@ export * from "./Table"
 export * from "./TextArea"
 export * from "./Button"
 export * from "./ErrorFallback"
+export * from "./HighlightText"
